+++ src/libs/ArrayExtended.ts
@@ -1,0 +1,9 @@
+import { ResultObject } from './Result/ResultObject';
+import { outputRealTimeResultString } from './ConsoleOut/RealTimeResult';
+
+export class ArrayWithRealTimeOutput extends Array<ResultObject> {
+    push(result: ResultObject) {
+        outputRealTimeResultString(result.resultCode);
+        return super.push(result);
+    }
+}
