+++ src/libs/config.ts
@@ -1,1 +1,3 @@
 export const VERSION = 'v0.0.1';
+export const OPENING_MESSAGE = `            Morse Test Runner ${VERSION}`;
+export const MORSE = '-- --- .-. ... .  - . ... -  .-. ..- -. -. . .-. ';
