+++ src/tests/morse.ts
@@ -54,8 +54,6 @@ const tests = {
             });
         })
     },
-
-
 };
 
 doTest(tests);
